requirements.txt
python-decouple
python-nmap